% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_media_info.R
\name{get_media_info}
\alias{get_media_info}
\title{Get media info in long data frame format}
\usage{
get_media_info(x)
}
\arguments{
\item{x}{directory with video files or vector of file paths}
}
\description{
None video files in x may cause errors.
}
