% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata_extended}
\alias{get_metadata_extended}
\title{Get all the meta data of (video) files}
\usage{
get_metadata_extended(x, metadata_tag = "comment")
}
\arguments{
\item{x}{file paths or directory}

\item{metadata_tag}{tag name}
}
\description{
Get all the meta data of (video) files
}
