% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata_tag}
\alias{get_metadata_tag}
\title{Get tag from a (video) file}
\usage{
get_metadata_tag(filepath, tag = "comment")
}
\arguments{
\item{filepath}{video file on disk}

\item{tag}{tag name}
}
\value{
tag value, character
}
\description{
Get tag from a (video) file
}
