% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_to_video.R
\name{img_to_video}
\alias{img_to_video}
\title{Create video from image sequence}
\usage{
img_to_video(
  x,
  out_path = NULL,
  out_name = "out",
  overwrite = F,
  fps_out = 30,
  scale = NULL,
  encoder = c("libx264", "libx265", "libvpx_vp9", "gif"),
  container = c("mp4", "mkv", "webm", "gif", "avi", "mov"),
  log_level = c("", "info", "quiet", "panic", "fatal", "error", "warning", "verbose",
    "debug", "trace"),
  quality_crf = NULL,
  run_cmd = T,
  flags_add_before_i = "",
  flags_add_after_i = "-pix_fmt yuv420p",
  which_ffmpeg = c("ffmpeg", "ffmpeg7")
)
}
\arguments{
\item{x}{path to image folder}

\item{out_path}{output folder path; if NULL: folder created next to x with
filename of x}

\item{out_name}{name of output file w/o file extension}

\item{overwrite}{overwrite files on disk? if FALSE and file exists in out_path,
file name is augmented by increasing integer at the end}

\item{fps_out}{-f flag: fps in output file}

\item{scale}{resize output images, c(width, height); either something like
c(1280, 720) for exact pixels or c(0.5, 0.5) for fraction of input}

\item{encoder}{video encoder/format; checked in combination with
container}

\item{container}{output container/ file ext, if NULL same as input file ext}

\item{log_level}{amount of log info}

\item{quality_crf}{-crf flag: see ?video_batch_convert}

\item{run_cmd}{run ffmpeg command (TRUE) or just return the string (FALSE)}

\item{flags_add_before_i}{additional flags for ffmpeg command before input file}

\item{flags_add_after_i}{additional flags for ffmpeg command after input file}

\item{which_ffmpeg}{which version to use}
}
\description{
Images need to be in one folder and named in sequential order
}
