% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_shquoted}
\alias{is_shquoted}
\title{Detect Shell-Quoted Strings}
\usage{
is_shquoted(x)
}
\arguments{
\item{x}{A character vector of strings to inspect.}
}
\value{
A logical vector with one element for each element of `x`. An element
  is `TRUE` when the corresponding string looks shell-quoted and `FALSE`
  otherwise. Missing inputs produce `NA`.
}
\description{
Determines whether strings appear to have been produced by `shQuote()`
using either POSIX-style single quoting or Windows-style double quoting.
}
\examples{
is_shquoted("'file name.txt'")
is_shquoted('"file name.txt"')
is_shquoted("file name.txt")

}
\seealso{
[shQuote()], [unshquote()]
}
\keyword{internal}
