% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{parse_cmd}
\alias{parse_cmd}
\title{Dissect ffmpeg command to its parts}
\usage{
parse_cmd(cmd)
}
\arguments{
\item{cmd}{ffmpeg command}
}
\value{
data frame
}
\description{
Dissect ffmpeg command to its parts
}
