% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_concat.R
\name{video_concat}
\alias{video_concat}
\title{Concatenate files}
\usage{
video_concat(x, out_path = NULL, out_name = NULL, overwrite = F)
}
\arguments{
\item{x}{directory with (video) files}

\item{out_path}{output path, defaults to x}

\item{out_name}{output name, defaults to 'firstfilename'___'lastfilename'}

\item{overwrite}{overwrite output file?}
}
\value{
nothing
}
\description{
Must be files of same type in folder x, named in sequential order.
}
