% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_to_video_estimate_result.R
\name{video_to_video_estimate_result}
\alias{video_to_video_estimate_result}
\title{Estimate filesize and conversion time by different settings}
\usage{
video_to_video_estimate_result(
  x,
  preset = c("ultrafast", "superfast", "veryfast", "faster", "fast", "medium", "slow",
    "slower", "veryslow"),
  quality_crf = 28,
  sample_number = 10,
  sample_duration = 10,
  video_to_video_args = list()
)
}
\arguments{
\item{x}{path to video file}

\item{preset}{preset flags to iterate}

\item{quality_crf}{crf flag to iterate}

\item{sample_number}{how many sample per crf/preset combination}

\item{sample_duration}{duration in sec of each sample}

\item{video_to_video_args}{further arguments to av2::video_to_video}
}
\value{
list
}
\description{
All combination of preset x quality_crf are tested
}
